Deleted emissions.css, tips.css and all references to them in the HTML
Because of my previous commit which consolidated the CSS into general.css, the tips and emissions CSS files are now redundant. I have deleted them and I also deleted the links to them in the HTML for those pages. I have a local copy of those CSS files in case we need to restore them.