Commented 2 lines in general.css
Commented two lines of css in general.css. First I commented the line that makes the whole page fade in, because I want to see if the site is better without it. Second, a line that appears to have typos in it, but I don't want to remove it entirely just in case.